apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "review-control-plane-api"
version = "0.1.0"
description = "FastAPI backend for 008 review control plane"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "websockets>=13.0",
  "python-docx>=1.1.0",
  "docx2txt>=0.8",
  "pdfplumber>=0.11.0",
  "python-multipart>=0.0.9",
  "pydantic>=2.8.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.23.0",
  "ruff>=0.6.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
testpaths = ["tests"]
